+++ src/cli.rs
@@ -51,12 +51,30 @@ pub fn get_app() -> App<'static, 'static> {
         .author("Logan Saso <logansaso+tech@gmail.com>")
         .about("A small program to create remote git repositories from the command line.")
         .version(env!("CARGO_PKG_VERSION"))
-        .arg(
-            Arg::with_name("endpoint")
-                .long("endpoint")
-                .takes_value(true),
-        )
         .subcommand(github::subcommand())
         .subcommand(gitlab::subcommand())
         .subcommand(bitbucket::subcommand())
+        .arg(
+            Arg::with_name("endpoint")
+                .long("endpoint")
+                .takes_value(true)
+                .help("Sets a custom endpoint to POST to, useful if you want a private instance and know the api matches one gitpub supports.")
+                .conflicts_with("set_remote"),
+        )
+        .arg(
+            Arg::with_name("set_remote")
+                .long("set_remote")
+                .help("Sets the remote of the local dir after successful creation."),
+        ).arg(
+            Arg::with_name("remote_name")
+                .long("remote_name")
+                .help("Designates a custom name for setting remote. Defaults to origin.")
+                .default_value("origin")
+                .hide_default_value(true),
+        ).arg(
+            Arg::with_name("ssh_remote_format")
+                .long("ssh_remote_format")
+                .help("Attempts to convert the git remote url into ssh format. If it fails (the provider doesn't support ssh format), the remote isn't set.")
+                .conflicts_with("endpoint"),
+        )
 }