+++ src/main.rs
@@ -2,7 +2,10 @@ use cli::Gitpo;
 use reqwest::StatusCode;
 use std::process::exit;
 mod cli;
+mod git;
 mod provider;
+
+use git::add_remote;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let matches = cli::get_app().get_matches();
@@ -26,7 +29,23 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     match status {
         StatusCode::OK | StatusCode::CREATED => {
             let apiloc = config.extract_url(&headers);
+            let (remote_url, can_use_ssh) = match (
+                config.ssh_url(&headers),
+                matches.is_present("ssh_remote_format"),
+            ) {
+                (Some(url), true) => (url, true),
+                _ => (config.extract_url(&headers), false),
+            };
             println!("Repo created: {}", apiloc);
+            let remote_name = matches
+                .value_of("remote_name")
+                .expect("This should default to origin, so something is wrong.");
+            if matches.is_present("set_remote")
+                && ((matches.is_present("ssh_remote_format") && can_use_ssh) || !can_use_ssh)
+                && !add_remote(remote_name, &remote_url)
+            {
+                eprintln!("Couldn't set remote.");
+            }
         }
         StatusCode::UNPROCESSABLE_ENTITY | StatusCode::BAD_REQUEST => {
             eprintln!("The provider had an issue processing this request. Perhaps the repository already exists, or you're using an unsupported option. e.g. Enabling projects on a repo in an org that has them disabled.");