+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## [0.4.0] - 2019-05-07
+## [0.4.1] - 2019-05-11
+### Removed
+
+* Removed `--endpoint` from bitbucket subcommand. It is taken care of via a global param now.
+
+## [0.4.0] - 2019-05-11
+### Added
+* Added `--ssh_remote_format`, `--set_remote`, and `--remote_name` commands. These are global commands and must be used before the subcommand. Example: `gitpub --ssh_remote_format github -n "name"` Check README for more info.
+
 ### Changed
 
 * This version encompasses changes to the CLI interface. It was re-written from structopt to use a clap app.